+++ test/utils/pageobjects/personalAccount.page.js
@@ -5,16 +5,8 @@ class PersonalAccountPage extends Page {
     return $('h3.text-align-left');
   }
 
-  get headerLogo() {
-    return $('div.header-logo > a');
-  }
-
   async validateUserLoginIsCompleted(expectedGreeting) {
     await expect(await this.greeting).toHaveText(expectedGreeting);
-  }
-
-  async goToHomePage() {
-    await (await this.headerLogo).click();
   }
 }
