+++ test/utils/waiters/wait-of-redirect.js
@@ -1,0 +1,3 @@
+export default function waitOfRedirect(partOfWaitingUrl) {
+   browser.waitUntil(() => browser.getUrl().then((pageUrl) => pageUrl.indexOf(partOfWaitingUrl) > -1), 5000);
+}
